Rename files more clearly, we can't use webpack.config.js because storybook will try to do some things automatically with that
